Updating based on first round of code review discussion
