+++ packages/http-client-java/generator/http-client-generator-core/src/main/java/com/microsoft/typespec/http/client/generator/core/mapper/ClientMethodMapper.java
@@ -22,17 +22,13 @@ import com.microsoft.typespec.http.client.generator.core.model.clientmodel.Imple
 import com.microsoft.typespec.http.client.generator.core.model.clientmodel.MethodPageDetails;
 import com.microsoft.typespec.http.client.generator.core.model.clientmodel.MethodParameter;
 import com.microsoft.typespec.http.client.generator.core.model.clientmodel.MethodPollingDetails;
-import com.microsoft.typespec.http.client.generator.core.model.clientmodel.ParameterTransformations;
 import com.microsoft.typespec.http.client.generator.core.model.clientmodel.ProxyMethod;
-import com.microsoft.typespec.http.client.generator.core.model.clientmodel.ProxyMethodParameter;
 import com.microsoft.typespec.http.client.generator.core.model.clientmodel.ReturnValue;
 import com.microsoft.typespec.http.client.generator.core.model.javamodel.JavaVisibility;
 import com.microsoft.typespec.http.client.generator.core.util.MethodNamer;
 import com.microsoft.typespec.http.client.generator.core.util.MethodUtil;
 import com.microsoft.typespec.http.client.generator.core.util.SchemaUtil;
 import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -145,13 +141,13 @@ public class ClientMethodMapper implements IMapper<Operation, List<ClientMethod>
         final List<ClientMethod> methods = new ArrayList<>();
 
         for (Request request : requests) {
-            List<ProxyMethod> proxyMethods = proxyMethodsMap.get(request);
+            final List<ProxyMethod> proxyMethods = proxyMethodsMap.get(request);
             for (ProxyMethod proxyMethod : proxyMethods) {
 
                 final List<Parameter> codeModelParameters
                     = getCodeModelParameters(request, operation, isProtocolMethod);
-                final ParameterProcessor.Result result = ParameterProcessor.process(request, codeModelParameters,
-                    proxyMethod.hasParameterOfType(ClassType.BINARY_DATA), isProtocolMethod);
+                final ClientMethodParameterProcessor.Result result = ClientMethodParameterProcessor.process(request,
+                    codeModelParameters, proxyMethod.hasParameterOfType(ClassType.BINARY_DATA), isProtocolMethod);
                 final MethodOverloadType defaultOverloadType = result.hasNonRequiredParameters
                     ? MethodOverloadType.OVERLOAD_MAXIMUM
                     : MethodOverloadType.OVERLOAD_MINIMUM_MAXIMUM;
@@ -162,7 +158,6 @@ public class ClientMethodMapper implements IMapper<Operation, List<ClientMethod>
                     .validateExpressions(result.validateParameterExpressions)
                     .parameterTransformations(result.parameterTransformations)
                     .methodVisibilityInWrapperClient(methodVisibilityInWrapperClient(operation, isProtocolMethod))
-                    .methodPageDetails(null)
                     .build();
 
                 final MethodNamer methodNamer
@@ -343,6 +338,39 @@ public class ClientMethodMapper implements IMapper<Operation, List<ClientMethod>
         }
     }
 
+    /**
+     * Gets the visibility for the client methods when generator configured to generate the wrapper clients.
+     * <p>
+     * The generation of the wrapper clients means separate clients are generated for async and sync APIs, such as
+     * FooAsyncClient and FooClient. These public clients are a thin layer over a shared service client implementation
+     * class (FooClientImpl), which is scoped to the implementation namespace and hidden from the user.
+     * </p>
+     * <p>This method returns the visibility of the client methods in the service client implementation class.</p>
+     * <p>
+     * In Data Place Generator (DPG) mode, wrapper clients are generated by default, DPG mode is enabled by --data-plane
+     * flag.
+     * Refer <a href="https://github.com/Azure/autorest.java?tab=readme-ov-file#minimal-data-plane-clients">Minimal
+     * Data-Plane Clients</a>.
+     * </p>
+     * <p>
+     * In vanilla mode, the wrapper clients are not generated by default, this mode is typically used when public-facing
+     * clients
+     * involve complex logic and therefore need to be handwritten. However, it is possible to configure wrapper client
+     * generation using the --generate-sync-async-clients and --generate-client-as-impl flags. For more details on these
+     * flags, refer to the
+     * <a href="https://github.com/Azure/autorest.java?tab=readme-ov-file#settings">Code generator settings</a> section.
+     * </p>
+     * <p>
+     * In management (ARM) modes - Lite and Premium, wrapper clients are not generated. The Lite mode only has Sync
+     * APIs,
+     * where it exposes service implementation client as public client. The Premium mode has both async and sync in
+     * one client (to ease the migration from Track-1 libraries with async and sync mixed in one client).
+     * </p>
+     *
+     * @param operation the operation.
+     * @param isProtocolMethod true if the DPG mode is enabled.
+     * @return the visibility of the client methods in the service client implementation class.
+     */
     private static JavaVisibility methodVisibilityInWrapperClient(Operation operation, boolean isProtocolMethod) {
         if (operation.getInternalApi() == Boolean.TRUE
             || (isProtocolMethod && operation.getGenerateProtocolApi() == Boolean.FALSE)) {
@@ -638,6 +666,10 @@ public class ClientMethodMapper implements IMapper<Operation, List<ClientMethod>
                 methodNamer.getLroModelBeginMethodName(), methods, createMethodArgs);
         }
 
+        // In Protocol mode (i.e., Data Place Generator (DPG)), the LRO begin method returns BinaryData -
+        // 1. PollerFlux<BinaryData, BinaryData> begin[operation]Async(..)
+        // 2. PollerFlux<BinaryData, Void> begin[operation]Async(..)
+        // Void for final result if the LRO is a resource delete operation.
         final ClientMethod lroBaseMethod = baseMethod.newBuilder()
             .methodPollingDetails(pollingMetadata.asMethodPollingDetailsForBinaryDataResult())
             .build();
@@ -908,24 +940,14 @@ public class ClientMethodMapper implements IMapper<Operation, List<ClientMethod>
     /**
      * Gets the Context parameter.
      *
-     * @param isProtocolMethod Whether the method is a protocol method.
+     * @param isProtocolMethod Whether the client method using the Context is a protocol method.
      * @return The Context parameter.
      */
     protected ClientMethodParameter getContextParameter(boolean isProtocolMethod) {
         if (isProtocolMethod) {
             return ClientMethodParameter.REQUEST_OPTIONS_PARAMETER;
         } else {
-            return new ClientMethodParameter.Builder().description("The context to associate with this operation.")
-                .wireType(ClassType.CONTEXT)
-                .name("context")
-                .requestParameterLocation(RequestParameterLocation.NONE)
-                .annotations(Collections.emptyList())
-                .constant(false)
-                .defaultValue(null)
-                .fromClient(false)
-                .finalParameter(false)
-                .required(false)
-                .build();
+            return ClientMethodParameter.CONTEXT_PARAMETER;
         }
     }
 
@@ -988,95 +1010,6 @@ public class ClientMethodMapper implements IMapper<Operation, List<ClientMethod>
             this.methodNamer = methodNamer;
             this.generateRequiredOnlyParametersOverload = settings.isRequiredParameterClientMethods()
                 && defaultOverloadType == MethodOverloadType.OVERLOAD_MAXIMUM;
-        }
-    }
-
-    private static final class ParameterProcessor {
-        static ParameterProcessor.Result process(Request request, List<Parameter> codeModelParameters,
-            boolean mapFluxByteBufferToBinaryData, boolean isProtocolMethod) {
-
-            final List<ClientMethodParameter> parameters = new ArrayList<>();
-            final List<String> requiredParameterExpressions = new ArrayList<>();
-            final Map<String, String> validateParameterExpressions = new HashMap<>();
-            final boolean isJsonPatch = MethodUtil.isContentTypeInRequest(request, "application/json-patch+json");
-
-            final ParametersTransformationProcessor transformationProcessor
-                = new ParametersTransformationProcessor(isProtocolMethod);
-            for (Parameter parameter : codeModelParameters) {
-                final ClientMethodParameter clientMethodParameter
-                    = toClientMethodParameter(parameter, isJsonPatch, mapFluxByteBufferToBinaryData, isProtocolMethod);
-                if (request.getSignatureParameters().contains(parameter)) {
-                    parameters.add(clientMethodParameter);
-                }
-                transformationProcessor.addParameter(clientMethodParameter, parameter);
-
-                if (!parameter.isConstant() && parameter.getGroupedBy() == null) {
-                    final MethodParameter methodParameter;
-                    final String expression;
-                    if (parameter.getImplementation() != Parameter.ImplementationLocation.CLIENT) {
-                        methodParameter = clientMethodParameter;
-                        expression = methodParameter.getName();
-                    } else {
-                        ProxyMethodParameter proxyParameter = Mappers.getProxyParameterMapper().map(parameter);
-                        methodParameter = proxyParameter;
-                        expression = proxyParameter.getParameterReference();
-                    }
-
-                    if (methodParameter.isRequired() && methodParameter.isReferenceClientType()) {
-                        requiredParameterExpressions.add(expression);
-                    }
-                    final String validation = methodParameter.getClientType().validate(expression);
-                    if (validation != null) {
-                        validateParameterExpressions.put(expression, validation);
-                    }
-                }
-            }
-            final ParameterTransformations parameterTransformations = transformationProcessor.process(request);
-
-            return new Result(parameters, requiredParameterExpressions, validateParameterExpressions,
-                parameterTransformations, hasNonRequiredParameters(parameters));
-        }
-
-        private static ClientMethodParameter toClientMethodParameter(Parameter parameter, boolean isJsonPatch,
-            boolean mapFluxByteBufferToBinaryData, boolean isProtocolMethod) {
-            final ClientMethodParameter clientMethodParameter;
-            if (isJsonPatch) {
-                clientMethodParameter = CustomClientParameterMapper.getInstance().map(parameter, isProtocolMethod);
-            } else {
-                clientMethodParameter = Mappers.getClientParameterMapper().map(parameter, isProtocolMethod);
-            }
-
-            if (mapFluxByteBufferToBinaryData
-                && clientMethodParameter.getClientType() == GenericType.FLUX_BYTE_BUFFER) {
-                return clientMethodParameter.newBuilder()
-                    .rawType(ClassType.BINARY_DATA)
-                    .wireType(ClassType.BINARY_DATA)
-                    .build();
-            } else {
-                return clientMethodParameter;
-            }
-        }
-
-        private static boolean hasNonRequiredParameters(List<ClientMethodParameter> parameters) {
-            return parameters.stream().anyMatch(p -> !p.isRequired() && !p.isConstant());
-        }
-
-        static final class Result {
-            private final List<ClientMethodParameter> parameters;
-            private final List<String> requiredParameterExpressions;
-            private final Map<String, String> validateParameterExpressions;
-            private final ParameterTransformations parameterTransformations;
-            private final boolean hasNonRequiredParameters;
-
-            private Result(List<ClientMethodParameter> parameters, List<String> requiredParameterExpressions,
-                Map<String, String> validateParameterExpressions, ParameterTransformations parameterTransformations,
-                boolean hasNonRequiredParameters) {
-                this.parameters = parameters;
-                this.requiredParameterExpressions = requiredParameterExpressions;
-                this.validateParameterExpressions = validateParameterExpressions;
-                this.parameterTransformations = parameterTransformations;
-                this.hasNonRequiredParameters = hasNonRequiredParameters;
-            }
         }
     }
 }
